backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
motor==3.4.0
pymongo[srv]==4.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
openai==1.30.1
pdfplumber==0.11.0
Pillow==10.3.0
httpx==0.27.0
